feat: enforce user uniqueness constraints and improve transaction logging consistency
- Added unique constraints for  Email and ContactNumber in the User entity and database schema
- Updated AuthService registration flow to validate duplicate email and contact number before account creation
- Introduced IsEmailTakenAsync and IsContactNumberTakenAsync in IAuthRepository and implemented repository-level checks
- Standardized TransactionType logging by storing enum values as integers across services and audit logs
- Updated AppDbContext configuration and migrations to support unique indexes and transaction type consistency
- Fixed expense transaction logging to correctly persist TransactionType values
- Corrected and standardized audit log messages for registration, authentication, and transaction activities
- Performed code cleanup and refactored related logic to improve data integrity and logging reliability

+++ BudgetWise.BLL/Services/ExpenseService.cs
@@ -70,7 +70,7 @@ namespace BudgetWise.BLL.Services
             Log.ForContext("IsAuditLog", true)
                .ForContext("UserId", userId)
                .ForContext("Username", user!.Username)
-               .ForContext("Type", TransactionType.Create)
+               .ForContext("Type", (int)TransactionType.Create)
                .ForContext("EntityName", "Expense")
                .ForContext("Activity", $"Created expense '{newExpense.Description}'.")
                .Information($"'{user.Username}' created expense '{newExpense.Description}'.");
@@ -98,7 +98,7 @@ namespace BudgetWise.BLL.Services
             Log.ForContext("IsAuditLog", true)
                .ForContext("UserId", userId)
                .ForContext("Username", user!.Username)
-               .ForContext("Type", TransactionType.Update)
+               .ForContext("Type", (int)TransactionType.Update)
                .ForContext("EntityName", "Expense")
                .ForContext("Activity", $"Updated expense '{existingExpense.Description}'.")
                .Information($"'{user.Username}' updated expense '{existingExpense.Description}'.");
@@ -120,7 +120,7 @@ namespace BudgetWise.BLL.Services
             Log.ForContext("IsAuditLog", true)
                .ForContext("UserId", userId)
                .ForContext("Username", user!.Username)
-               .ForContext("Type", TransactionType.Delete)
+               .ForContext("Type", (int)TransactionType.Delete)
                .ForContext("EntityName", "Expense")
                .ForContext("Activity", $"Deleted expense '{existingExpense.Description}'.")
                .Information($"'{user.Username}' deleted expense '{existingExpense.Description}'.");
